Optional enum fields: tweak accessor method
This changes the generated accessor methods on optional enum fields to
return the default enum value, instead of an Option containing the enum
value. This is important because default values for enums are set on a
per-field basis, not on the enum type itself. Without this change there's
no way to get the default value of the enum out. It's also more ergonomic
when enums are being used correctly (the first value is a true default,
or an 'unknown' variant).