message-graph: dont consider messages to be nested when repeated
This is a workaround for handling nested oneof fields appropriately.
MessageGraph is changed to not consider messages to be nested when they
are nested behind a repeated field. This will make MessageGraph a little
bit more difficult to reuse for Eq/Ord, but we'll cross that bridge when
we get there.