prost-build: pass Config, ServiceGenerator by mut ref
Changing ServiceGenerator::{generate, finalize} to take `&mut self` is
motivated by tower-rs/tower-grpc#21. The `ServiceGenerator` change
necessitated changing Config::compile_protos to take `&mut self` as
well. I don't expect the latter change to impact many people, since it's
the more advanced API (as opposed to the standalone `compile_protos`
function), and to be useful the API required a mutable config, anyway.
It also provides flexibility in the future if we want to do something
else requiring mutable access to the config.

This is a breaking change.