prost-build: strip enum name from enum variant prefix
This commit modifies prost-build to strip the enum name from enum
variants which include the enum name as a prefix. This style is common
with Protobuf (since it's heavily tied to C++ name resolution), however
it's non-idiomatic in Rust. To restore the previous behavior, a new
Config option is provided: `retain_enum_prefix`. This is a breaking
change.