prost_build::Config::extern_path
This commit introduces a new API on prost-build that allows Protobuf
packages and types to be declared as extern, along with a Rust path to
the corresponding pre-generated types. See the method API docs for
details.

The `extern_paths` API replaces and generalizes the `map_types` API
contributed by Eyal Kalderon and Trinh Hoang Anh in #111.