Add support for `deprecated` attributes on fields.
Protobufs support `deprecated` field options. These field options allow
the field to continue to be supported, but allow for language specific
code generators to include that the field has been deprecated in the
generated code. As discussed in
https://github.com/danburkert/prost/issues/221 we should consider
supporting this in `prost`.

This commit adds support for this feature in `prost-build` by checking
the field options, and if present emitting a `#[deprecated]` attribute.

One short coming of this change is that it is difficult to test that the
`deprecated` attribute is correctly added with an assertion since the
attribute is not accessible at runtime.