Add protoc plugin handling to the prost-build Config and basic plugin
This refactors some of the private logic within prost-build's lib.rs so
that it is compatible with generating the logic necessary to run a
plugin for protoc or a build.rs file.  In both cases, the output can be
changed based on how the Config is built.  This adds the capability for
paramters passed in via protoc to a plugin to allow adjust this
configuration, but the mechanics of that are left as a TODO until the
design is solidified.