CI: disable Cargo target directory caching (#322)
* Update formatting for rustfmt 1.4.12-stable

The handling of rustfmt for the varint! macro changes between rustfmt
1.4.11-stable and rustfmt 1.4.12-stable.  The new formatting is
acceptable to both versions.

* Remove cache action step that is causing problems for CI

I've noticed a number of github actions reporting failing or cancelled
on the cache step.  This removes them, for now, to provide a better
signal for the project, until why the cache action fails can be
determined.

Builds on PR #312.

* remove all caching

Co-authored-by: David Freese <dvfreese@gmail.com>