include_file_descriptor_set -> file_descriptor_set_path
This tweaks the file descriptor set option to take a path argument
instead of a fixed path. This allows it to work with crates who invoke
`prost-build` multiple times.  It also saves `prost-build` from having
to read and re-encode the file descriptor set, which could have subtle
effects on it.