prost-build: internal refactor with PathMap
This commit adds a new internal data structure in prost-build: the
`PathMap`. `PathMap` better supports the frequent configuration pattern
of associating fully-qualified Protobuf paths with a value. The
fully-qualified paths are matched according to the prefix/suffix rules
already used by configuration options such as `Config::btree()`.
`PathMap` will be a good basis to build further support for custom
string, bytes, and maps types in the future, in addition to other code
generation options.

This commit should not have any publicly visible changes, except for a
small tweak to allow for more intuitive field matching when the pattern
is a suffix-matched message.