Unhide `use std::io::Result` in doc example; it's important
Copying this example from docs.rs/prost-build into your own build.rs
doesn't compile because `std::result::Result` needs 2 type args and this
only has 1.

Because this is in a doc comment rather than rustdoc, you can't just
click on the `Result` in the HTML to see which `Result` it is.

So instead, expose the `use` to let people get started more easily.