Avoid double-copy merging String from non-Bytes input.
The optimization in https://github.com/tokio-rs/prost/pull/449 avoids
copies when both the input and output are `Bytes`, but it can cause
double copies when neither the input nor output is `Bytes`.

This is easy to trigger by using `&[u8]` as the input buffer and merging
a `Vec<u8>` or `String`, both of which are generated by `prost-build`.

Introduce `encoding::bytes::merge_one_copy()` and use it to guarantee
that `encoding::string::merge()` copies exactly once. The new function
is only visible to the encoding module so it does not affect the API.

It would also be useful to make `prost-derive` call this for
`#[prost(bytes="vec")]`, but that may require making it public, which
is a more drastic and less justified change. It may be better to wait
for type specialization to be stabilized.