Use `rsplitn` and `splitn` but add a clippy allowance
It turns out you get a compilation error on Rust 1.51.0 if you use the `split_once` and `rsplit_once` methods since they were unstable at that version.

In Rust stable (currently 1.58.1), there are clippy recommendations that
say to use those methods instead of `rsplitn` and `splitn` if you're only splitting once.

I figured the best way to handle this since we're targeting 1.51 in addition to stable is to just set specific ignore in that case so the old version still compiles successfully.