Make prost-build process code blocks in comments
Because Google make no statement about the format of the content
of comments in protocol buffers, since they're essentially just
comments, there is a mix of content out there, some markdown, some
not.  And where it parses as markdown, some comments have code
blocks in them, for example `google.api.HttpRule` does.

Since prost-build puts these comments into the output code as
documentation comments, we treat the text as markdown and process
it to ensure that code blocks are not incorrectly treated as
Rust doctests.

Signed-off-by: Daniel Silverstone <dsilvers@digital-scurf.org>