(prost-build): Make markdown cleanup optional (off by default)
In order to ensure that we do not end up with extra dependencies
being added to everyone's build, make the cleanup feature optional
and not enabled by default.  This will permit people to opt in to
the feature.  If over time it is discovered that enough people
opt in, we can make it default later.

Signed-off-by: Daniel Silverstone <dsilvers@digital-scurf.org>