fix(build): Don't add extra space before RustDoc comments (#694)
In v0.11.0 the `sanitize_line` function was updated to add a space to
the beginning of non-empty RustDoc comments, but often those comments
already contain a space. This commit updates the check to ensure the
extra space is only added if it doesn't already exist.

We also need to check that the line doesn't start with _multiple_
spaces, as in the case of multi-line list items in Markdown docs. These
items _will_ need to be indented to match their list marker's
indentation.

Note that this commit actually modifies the `protobuf.rs` file as well
because code blocks inside RustDoc comments were previously
over-indented.

Fixes #693.