feat(prost-build): optionally generate builder API
For each message, optionally generate a struct to provide the
forward-compatible builder API, with inherent methods
for field setters and the .build() method to produce the fully
initialized message struct.

If generation of builders is enabled with a name `Builder` for a message
named `Foo`, an impl block with the associated fn `Foo::builder()` is
appended to the top-level file scope, and the definition of the builder
struct `Builder` and its methods is appended to the module `foo`.

Reuse the helpers that generate field definitions for a message,
but use a mode to output code for the builder fields,
which must be private and devoid of docs and prost annotations.

The added Config::builder method populates a PathMap
with values given in the second argument to configure the builder name.