feat: OpenEnum, Config::typed_enum_fields
Introduce OpenEnum to represent fields of Protobuf enumerated types
with the possibility of unknown values.

A new enum_type annotation is supported in the prost attribute
inside derives, which allows to specify the type-checked representation
of enum types in message fields and oneof variants.
The accepted values are "open" or "closed".

Use OpenEnum in generated code for enum-typed fields of messages
instead of i32, if the `enum_type` annotation is set to "open".

OpenEnum features get and set methods, which are meant to be
the primary ways to get and set open enumeration fields,
instead of the getters and setters generated by Message derive.

Add typed_enum_fields method to prost-build configuration, which
allows type-checked representation of enumerations in fields of
message structs and variants of oneof enums. The argument and the
invocation order works like with the boxed method.

Depending on the syntax (and preparing for the future support of
editions), the type-checked representation can be closed (for proto2)
or open (for proto3). The former is represented by the generated
enum type itself, while the latter is represented by OpenEnum
wrapping the enum type.