prost-build: refactor code generator
Collect message field data into structs that can be reused in multiple
passes. This will be useful for the upcoming generation of builders.

Add prost_path helper function to DRY repetitive expressions obtaining
the path to the prost crate.

Add boxed helper method to capture the logic of deciding whether a field
is boxed.