build: Ignore missing features in `tests` crates
Only the crate `tests-2015` uses the feature `edition-2015`, but the code is shared with the `tests` crate.

Same for `tests-no-std` which disables feature `std`, but reuses the code with the `tests` crate.

Therefore the `tests`, `tests-2018` and `tests-no-std` crates should be configured to ignore the lint about missing features.

Cargo clippy reports:
```
warning: unexpected `cfg` condition value: `edition-2015`
 --> tests-no-std/../tests/src/build.rs:5:14
  |
5 |     if #[cfg(feature = "edition-2015")] {
  |              ^^^^^^^-----------------
  |              |
  |              help: remove the condition
  |
  = note: no expected values for `feature`
  = help: consider adding `edition-2015` as a feature in `Cargo.toml`
  = note: see <https://doc.rust-lang.org/nightly/rustc/check-cfg/cargo-specifics.html> for more information about checking conditional configuration
  = note: `#[warn(unexpected_cfgs)]` on by default
```