style: Move encoding functions into separate modules (#1111)
* style: Move varint encoding to separate module

While the `encoding` module is undocumented and not stable, there are known users. Leave a alias in place to prevent breaking them.

* style: Move length delimiter encoding to separate module

Leave a alias in `lib.rs` to prevent a breaking change.

* style: Move wire type to separate module

While the `encoding` module is undocumented and not stable, there are known users. Leave a alias in place to prevent breaking them.