build(protobuf): Don't install unused test proto
The build script of crate `protobuf` installs some proto-files for testing in the prefix dir. However the only user is the same build script. Therefore we can skip the install and use the files directly.

Use `absolute` path instead of `canonical` path, as `protoc` on Windows is not able to process `canonical` paths. This is sort of documented in https://doc.rust-lang.org/std/fs/fn.canonicalize.html