ci: Run `cargo test` without `all-targets` (#1118)
Running `cargo test --all-targets` will not run doctests, as noted in the removed comments.

Another effect is that Cargo.toml instructions for not running a test are ignored. For example, benchmarks are tested, but no unit tests are inside and bins marked with `test = false` are run.

I diffed the text output of running `cargo test` and `cargo test --all-targets` and concluded that no useful test will be skipped when running just `cargo test`.