fix(prost-build): Resolve OneOf type name conflict with embedded message (#1294)
When a oneof field and a message have the same name, then `prost-build` will generate invalid code. The generated types have the same name. With this change the oneof enum it renamed to resolve the conflict. This is not a breaking change, because the old code didn't compile.

Example of generated code for `oneof_name_conflict.proto`:
```rust
// This file is @generated by prost-build.
#[derive(Clone, PartialEq, Eq, Hash, ::prost::Message)]
pub struct Bakery {
    #[prost(oneof = "bakery::BreadOneOf", tags = "1, 2")]
    pub bread: ::core::option::Option<bakery::BreadOneOf>,
}
/// Nested message and enum types in `Bakery`.
pub mod bakery {
    #[derive(Clone, Copy, PartialEq, Eq, Hash, ::prost::Message)]
    pub struct Bread {
        #[prost(int32, tag = "3")]
        pub weight: i32,
    }
    #[derive(Clone, PartialEq, Eq, Hash, ::prost::Oneof)]
    pub enum BreadOneOf {
        #[prost(string, tag = "1")]
        A(::prost::alloc::string::String),
        #[prost(message, tag = "2")]
        B(Bread),
    }
}
```