prost-derive: Add #[(prost(prost_path = "::prost")]
Add a new proc_macro attribute in `prost-derive`, which is used to let
the macro implementation know about the value of
`prost_build::Config::prost_path`.

As a result, the generated types no longer have `::prost` hardcoded in
them, and a re-exported prost crate can be used instead along with
generated code.

Fixes #939