feat: Configure prost path via `prost_build::Config` or `#[(prost(prost_path = "::prost")]` (#1274)
* prost-derive: Add #[(prost(prost_path = "::prost")]

Add a new proc_macro attribute in `prost-derive`, which is used to let
the macro implementation know about the value of
`prost_build::Config::prost_path`.

As a result, the generated types no longer have `::prost` hardcoded in
them, and a re-exported prost crate can be used instead along with
generated code.

Fixes #939

* Make path to ::prost_types configurable

Similarly to the previous commit, this allows customization of the path
to `prost_types` crate in case a re-export is desired. This can now be
done by calling the `prost_build::Config::prost_types_path` method.

* Add tests

* Address PR comments