feature: expose public error API for decode errors
This commit updates the Error API and exposes it publicly.

The location of the error becomes fully accessible to consumer code through the `ErrorPath` abstraction. It allows to read the logical path without exposing implementation details.

The error kind is turned into an enum of source errors. Each source error and the error kind itself are non-exhaustive. This allows to new fields in the future without breaking backwards compatibility.

Public constructors are provided to enable advanced use-cases in third-party crate that want to create new error types.

`DecodeError` uses the `core::error::Error` API to expose source errors. In particular, it does not print the source error in its own display message. This allows better integration with error reporting frameworks.